refactor(#290): split AppSidebar.tsx into sub-components
Extract SessionItem, ProjectsSection, and date-utils into separate
files under components/sidebar/. AppSidebar.tsx is now ~233 lines
(down from 795).
